+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * feat(`PageableInterface`): add `$start` parameter to `getPages()` method to
   ease batch resuming.
 * refactor: move common indexBy logic to separate package
+* feat(`SelectableAdapter`): add `indexBy` parameter
 
 # 0.11.2